once: replace Once with an Once based on the high-level Mutex
